--- MileStone_2/MiniTalk/Version3/Libft/README.md
@@ -1,34 +1,0 @@
-# LIBFT
-
-<div>
-
-  ![Project](https://img.shields.io/badge/Project-Libft-blue)
-  ![Version](https://img.shields.io/badge/Version-1.0-green)
-
-</div>
-
-The aim of this project is to build a personal C library, designed to solidify foundational knowledge in:
-- Core functions
-- Dynamic memory management
-- Makefiles
-- Libraries
-- Structs and linked lists
-
-These functions will become valuable tools throughout the course, with the potential to expand by adding more as needed.
-
-A Makefile should define the compilation rules, simplify certain flags, and include the following mandatory targets:
-- $(NAME)
-- all
-- clean
-- fclean
-
-The project should avoid unnecessary recompilation, meaning that running make consecutively should only return a message indicating that there’s nothing to be done.
-
----
-
-## Recommendations
-
-The recommended time frame for completing this project ranges from one week to two months, with one month being the average. 
-Do not try to rush too much, it's best to go at your own pace, develop a `habit`, and `learn`.
-If you get stuck on a function, it's advisable to `move on to the next one` and come back later. 
-Doing the bonus is highly recommended, as you will need and use it later.
