+++ MileStone_2/MiniTalk/Version3/server/server.c
@@ -6,7 +6,7 @@
 /*   By: yurolive <yurolive@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/28 16:39:14 by yurolive          #+#    #+#             */
-/*   Updated: 2024/10/28 16:39:14 by yurolive         ###   ########.fr       */
+/*   Updated: 2024/10/29 17:48:08 by yurolive         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
